Extract UpgradePlanAsync to command
This is a pure lift & shift with no refactors
